refactor: migrate from Netlify to Vercel
Move serverless function to Vercel api/ format, update fetch path,
and remove Netlify config files.
