--- src/admin-app/netlify.toml
@@ -1,10 +1,0 @@
-[build]
-  base = "src/admin-app"
-  command = "npm run build"
-  publish = "dist"
-  functions = "netlify/functions"
-
-[[redirects]]
-  from = "/*"
-  to = "/index.html"
-  status = 200
